Moved files to be more consistent.
