+++ src/components/crown/crownConfig/crownConfig.vue
@@ -56,8 +56,8 @@ import DeleteButton from '@/components/crown/crownButtons/deleteButton';
 import MessageFlowButton from '@/components/crown/crownButtons/messageFlowButton';
 import SequenceFlowButton from '@/components/crown/crownButtons/sequenceFlowButton';
 import AssociationFlowButton from '@/components/crown/crownButtons/associationFlowButton';
-import CrownDropdown from '@/components/crown/crownDropdown';
-import CrownBoundaryEventDropdown from '@/components/crown/crownBoundaryEventDropdown';
+import CrownDropdown from '@/components/crown/crownButtons/crownDropdown';
+import CrownBoundaryEventDropdown from '@/components/crown/crownButtons/crownBoundaryEventDropdown';
 import poolLaneCrownConfig from '@/mixins/poolLaneCrownConfig';
 import pull from 'lodash/pull';
 import store from '@/store';
